requirements.txt
Django==4.2.2
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0
whitenoise==6.5.0
djangorestframework==3.14.0